Move attention-focus code
