+++ opencog/guile/SchemeSmob.h
@@ -187,13 +187,6 @@ private:
 	static SCM ss_av_p(SCM);
 	static SCM ss_av_get_value(SCM);
 
-	// AttentionalFocus and AttentionalFocus Boundary
-	// XXX FIXME these should move to the attention bank!
-	static SCM ss_af(SCM n);
-	static SCM ss_af_size(void);
-	static SCM ss_set_af_size(SCM);
-	static SCM ss_stimulate(SCM, SCM);
-
 	// Free variables
 	static SCM ss_get_free_variables(SCM);
 	static SCM ss_is_closed(SCM);
